Enum based Singleton
The enum automatically guarantees the uniqueness of the singleton in all situations (it is also thread-safe).

But you can't persists fields data for an Enum object: only the literal value is persisted.

Another limitation is that you cannot inherit from an Enum. That would limit testing capabilities. 

If you are Ok with these limitations, you can use an Enum to create your Singleton classes.