Monostate: A pernitious variation of the Singleton.
The monostate manteins a single state, but fails in communicating the user that this is actually the case. The user will user the class as a normal multi-instances class, creating severall or many instances, and all of them will have the same state, changing over and over, whenever a setter gets called.

This is indeed a problem and this approach should never be used.