LineToPointAdapter
Assuming we only have a draw(Point) interface, if we need to draw lines (as in vectorial objects), we can write an adapter to pass from lines to points, as shown in the example.

With the current (noCaching) implementation, if we call the draw() method more than once, we generate the same points over and over again. It is so unefficient.